automatically migrates override config format
